requirements.txt
# continuum
hydra-core==1.2.0
# numpy
oauthlib==3.2.1
omegaconf==2.2.3
open-clip-torch==1.3.0
# pandas
# Pillow==9.2.0
pipreqs==0.4.11
# scikit-image==0.19.3
# scikit-learn==1.1.1
# scipy
tensorboard==2.10.0
timm==0.4.9
# tokenizers==0.12.1
# tqdm==4.64.0
# transformers==4.21.1
ftfy 
# regex

